otel-playground/requirements.txt
opentelemetry-distro
opentelemetry-exporter-otlp

opentelemetry-instrumentation-logging

# output of opentelemetry-bootstrap -a requirements (without pinned versions)
opentelemetry-instrumentation-aws-lambda
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-sqlite3
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-boto3sqs
opentelemetry-instrumentation-botocore
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-jinja2
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-tortoiseorm
opentelemetry-instrumentation-urllib3